requirements.txt
streamlit
torch
torchvision
timm
xgboost
opencv-python
numpy
scikit-learn
Pillow
reportlab
bcrypt